+++ source/Private/Write-AuditLog.ps1
@@ -1,5 +1,4 @@
-﻿function Write-AuditLog {
-    <#
+﻿<#
     .SYNOPSIS
         Writes log messages to the console and updates the script-wide log variable.
     .DESCRIPTION
@@ -49,12 +48,13 @@
 
         Sets the message to "End [FunctionName] log.", where FunctionName is the name of the calling function, and adds it to the log variable.
     .EXAMPLE
-        Write-AuditLog -End -OutputPath "C:\Logs\auditlog.csv"
+        Write-AuditLog -End -OutputPath "C:\Logs\auditLog.csv"
 
         Sets the message to "End Log", adds it to the log variable, and exports the log to a CSV file.
     .NOTES
     Author: DrIOSx
 #>
+function Write-AuditLog {
     [CmdletBinding(DefaultParameterSetName = 'Default')]
     param(
         ###
@@ -109,16 +109,26 @@
         [string]$OutputPath
     )
     begin {
-        $ErrorActionPreference = "SilentlyContinue"
+        $ErrorActionPreference = 'SilentlyContinue'
         # Define variables to hold information about the command that was invoked.
         $ModuleName = $Script:MyInvocation.MyCommand.Name -replace '\..*'
-        $FuncName = (Get-PSCallStack)[1].Command
+        $callStack = Get-PSCallStack
+        if ($callStack.Count -gt 1) {
+            $FuncName = $callStack[1].Command
+        }
+        else {
+            $FuncName = 'DirectCall'  # Or any other default name you prefer
+        }
+        #Write-Verbose "Funcname Name is $FuncName!" -Verbose
         $ModuleVer = $MyInvocation.MyCommand.Version.ToString()
         # Set the error action preference to continue.
-        $ErrorActionPreference = "Continue"
+        $ErrorActionPreference = 'Continue'
     }
     process {
         try {
+            if (-not $Start -and -not (Test-Path variable:script:LogString)) {
+                throw "The logging variable is not initialized. Please call Write-AuditLog with the -Start switch or ensure $script:LogString is set."
+            }
             $Function = $($FuncName + '.v' + $ModuleVer)
             if ($Start) {
                 $script:LogString = @()
@@ -169,19 +179,18 @@
             switch ($Severity) {
                 'Warning' {
                     Write-Warning ('[WARNING] ! ' + $Message)
-                    $UserInput = Read-Host "Warning encountered! Do you want to continue? (Y/N)"
+                    $UserInput = Read-Host 'Warning encountered! Do you want to continue? (Y/N)'
                     if ($UserInput -eq 'N') {
-                        Write-Output "Script execution stopped by user!"
-                        exit
+                        throw 'Script execution stopped by user.'
                     }
                 }
-                'Error'       { Write-Error ('[ERROR] X - ' + $FuncName + ' ' + $Message) -ErrorAction Continue }
-                'Verbose'     { Write-Verbose ('[VERBOSE] ~ ' + $Message) }
-                Default { Write-Information ('[INFORMATION] * ' + $Message)  -InformationAction Continue}
+                'Error' { Write-Error ('[ERROR] X - ' + $FuncName + ' ' + $Message) -ErrorAction Continue }
+                'Verbose' { Write-Verbose ('[VERBOSE] ~ ' + $Message) }
+                Default { Write-Information ('[INFO] * ' + $Message) -InformationAction Continue }
             }
         }
         catch {
-            throw "Write-AuditLog encountered an error (process block): $($_.Exception.Message)"
+            throw "Write-AuditLog encountered an error (process block): $($_)"
         }
 
     }
@@ -189,11 +198,11 @@
         try {
             if ($End) {
                 if (-not [string]::IsNullOrEmpty($OutputPath)) {
-                    $script:LogString | Export-Csv -Path $OutputPath -NoTypeInformation -Encoding utf8
+                    $script:LogString | Export-Csv -Path $OutputPath -NoTypeInformation
                     Write-Verbose "LogPath: $(Split-Path -Path $OutputPath -Parent)"
                 }
                 else {
-                    throw "OutputPath is not specified for End action."
+                    throw 'OutputPath is not specified for End action.'
                 }
             }
         }