+++ CHANGELOG.md
@@ -7,8 +7,25 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Add support for newer versions of each module.
+- Add support for numbers in pattern for app prefix.
+
+### Fixed
+
+- Fixed approved verb for main public function.
+- Some Error handling improvements.
+- Function names are now more consistent with approved verbs.
+- Refactored code to improve readability.
+- Consolidated functions to reduce complexity.
+- Minor Change to README.md.
+
+## [0.1.0] - 2023-07-15
+
+### Added
+
 - Add support for multiple attachments
 - Release to public.
+
 ## [0.1.0-preview0001] - 2023-07-15
 
 ### Added